refactor(LIB-1911): move FzInternalStepStatus and StepMeta types to types.ts
